backend/requirements.txt
boto3>=1.34.0
PyJWT>=2.8.0
jsonschema>=4.21.0
python-dotenv>=1.0.1
